feat: move HOCs to hooks
BREAKING CHANGE: removed all HOCs in favour of hooks
