+++ src/index.ts
@@ -1,7 +1,5 @@
 import BackgroundLayerChooser from './BackgroundLayerChooser/BackgroundLayerChooser';
 import BackgroundLayerPreview from './BackgroundLayerPreview/BackgroundLayerPreview';
-import onDropAware from './HigherOrderComponent/DropTargetMap/DropTargetMap';
-import timeLayerAware from './HigherOrderComponent/TimeLayerAware/TimeLayerAware';
 import {GeoLocation, useGeoLocation } from './hooks/useGeoLocation/useGeoLocation';
 import useMap from './hooks/useMap/useMap';
 import FloatingMapLogo from './Map/FloatingMapLogo/FloatingMapLogo';
@@ -22,9 +20,8 @@ export {
   InkmapPrintSpec,
   isWmsLayer,
   MapComponent,
-  onDropAware,
   PrintUtil,
-  timeLayerAware,
   useGeoLocation,
   useMap,
-  WmsLayer};
+  WmsLayer
+};
